Rework navbar again
When the user isn't logged in, just show the logo centered without any
links since they'll either be redirected to the login page or be in some
sort of dialog. When the user is logged in, only show the logout link in
the navbar since everything else is reachable from the home page. It
might need a bit more work particularly in the case of the OIDC consent
dialog where the user might be logged in here but we want more of a
"dialog" experience.