settings: Manage vault secrets
Rather than doing this only in the entrypoint script, handle it in the
settings utilities with hvac. The entrypoint script can still be used to
login to vault in an arbitrary way. Authentication to vault must be done
outside of eos-idp. If a token can't be found from the VAULT_TOKEN
environment variable or in the ~/.vault-token file, then the client
won't be connected and secrets won't be gathered from vault.

Various methods are provided mirroring the environment variable handling
so that settings can be managed in a uniform and flexible way.