main: Add ensuresuperuser management command
This command can be used to idempotently create a super user
non-interactively. It wraps createsuperuser and tries that first. If
there is an error, it tries to update an existing super user's fields.
The use case is to allow calling admin-setup multiple times without
triggering an error if the super user already exists.