Don't print vault token to console
Pass -no-print to vault login so it doesn't print out the token to the
console. When running the container this output would be logged, and
while the token will expire eventually, it could be hijacked. We only
rely on the stored token in ~/.vault-token, so the printed token is of
no use, anyways.