Use pipenv to manage dependencies
Many of the dependencies in the current Docker image are out of date and
there's currently no automated process to know when they should be
updated. Using `pipenv` provides 2 major benefits:

* The loose version requirements that were previously in
  `requirements.txt` can be kept while maintaining a list of pinned
  versions. This is helpful for deployments as the same python packages
  will always be installed without having to manage pinned versions.

* GitHub's dependabot will alert when the pinned versions in
  `Pipfile.lock` have security vulnerabilities. That will prompt someone
  to handle the issue and generate a new Docker image without the
  vulnerable components.

The `Pipfile` was imported from the existing `requirements.txt` file
using `pipenv install -r requirements.txt`. The `requirements.txt` file
was then regenerated using `pipenv lock -r > requirements.txt`.

As in the `Dockerfile`, the python version is specified to be 3.9.
Hopefully this isn't a burden, but it ensures that a locally tested
version matches what will go in the Docker image.

https://phabricator.endlessm.com/T33031