Specify DEFAULT_AUTO_FIELD for Django 3.2
Django 3.2 added a setting to handle the type for models that have an
automatic primary key. Setting this to `django.db.models.AutoField`
maintains the previous internal default and avoids getting a bunch of
migrations when they later change the internal default to
`django.db.models.BigAutoField`. This suppresses a bunch of warnings
whenever Django is started.

https://docs.djangoproject.com/en/3.2/ref/settings/#std:setting-DEFAULT_AUTO_FIELD