settings: Bump Facebook Graph API version to v19.0
Per the Graph API changelog[1], v19.0 is the current version. The only
Graph API used directly by allauth is the User Picture API[2], and that
does not appear to have changed. The actual login calls are managed by
the Facebook JS SDK, which presumably knows how to use the newer Graph
API version.

1. https://developers.facebook.com/docs/graph-api/changelog/
2. https://developers.facebook.com/docs/graph-api/reference/user/picture/

https://phabricator.endlessm.com/T35222