settings: Make email addresses optional for social accounts
For normal username/password accounts, it makes sense to require email
verification so that you can't claim someone else's address. It's also
important for resetting the user's password as that's the only means of
authentication.

However, if a user is signing up with a social account provider (Google
or Facebook), the provider is supplying the email address and we can
assume they've done some diligence that the user owns it. Furthermore,
they don't need the email address for password resets since
authentication comes from the provider. Allow email addresses to be
optional in that case.

https://phabricator.endlessm.com/T35763